1.增加map的部分接口 2.增加map的单元测试代码 3.修正rbtree的find函数中的一处错误 4.修正string的format方法中的错误
git-svn-id: http://qlanguage.googlecode.com/svn/trunk@11 6d982bb3-a5a7-393d-ffc5-582415be9c85

+++ QCore/Library/memory.h
@@ -100,7 +100,7 @@ namespace QLanguage
                     obj* current = chunk_list[i];
                     while(current)
                     {
-                        fprintf(fp, "0x%08X\n", current);
+                        fprintf(fp, "0x%08X\n", (int)current);
                         current = current->next;
                     }
                     fprintf(fp, "\n");
